fix few things according to Diego's review
