+++ src/utils/carousel.ts
@@ -1,4 +1,4 @@
-
+import { Event } from "api/event"
 /**
  * enum to differentiate the different types of widget we could have
  * different widget types have different background images
@@ -38,8 +38,14 @@ export interface CarouselItem {
     room?: string
 }
 
-export function formatTitle(str: string) {
-    return str
+/**
+ * Function that helps to format the title of an Highlight. I decided to transform the entire title to lowercase
+ * except for the first letter of each word with more than 3 characters.
+ * @param title
+ * @returns the transformed title
+ */
+export function formatTitle(title: string) {
+    return title
         .split(" ")
         .map(item =>
             item.length > 3
@@ -47,4 +53,61 @@ export function formatTitle(str: string) {
                 : item.toLowerCase()
         )
         .join(" ")
+}
+
+/**
+ * Function that allows to check the event_type of an event. So far, we filter the events that are lectures, exams or deadlines.
+ * @param typeId of the event
+ * @returns true/false
+ */
+export function checkEventType(typeId: number) {
+    return (
+        typeId === WidgetType.LECTURES ||
+        WidgetType.EXAMS ||
+        WidgetType.DEADLINE
+    )
+}
+
+export function createWidget(event: Event) {
+    const days = [
+        "Lunedi",
+        "Martedi",
+        "Mercoledi",
+        "Giovedi",
+        "Venerdi",
+        "Sabato",
+        "Domenica",
+    ]
+    const months = [
+        "Gennaio",
+        "Febbraio",
+        "Marzo",
+        "Aprile",
+        "Maggio",
+        "Giugno",
+        "Luglio",
+        "Agosto",
+        "Settembre",
+        "Ottobre",
+        "Novembre",
+        "Dicembre",
+    ]
+    const dateObj = new Date(event.date_start)
+    const resultDate =
+        days[(dateObj.getDay() - 1 + 7) % 7] +
+        " " +
+        dateObj.getDate().toString().padStart(2, "0") +
+        " " +
+        months[dateObj.getMonth()] +
+        " " +
+        dateObj.getFullYear()
+    const nextEvent: CarouselItem = {
+        id: event.event_id,
+        type: event.event_type.typeId,
+        date: resultDate,
+        time: event.date_start.toString().slice(11, 16),
+        title: event.title.it,
+        room: event.room?.acronym_dn,
+    }
+    return nextEvent
 }
